feat(nubank-auth): moving code to a nubank module
